Separate out auto-opponent into own folder
