+++ packages/rps/src/redux/auto-opponent/index.ts
@@ -1,0 +1,2 @@
+export { autoOpponent } from './auto-opponent';
+export { autoPlayer } from './auto-player';
